refactor(fuse): extract metadata.rs and fs.rs from lib.rs
- Move encrypt_metadata_to_json, merge_folder_children, spawn_metadata_publish,
  spawn_bin_entry_publish, spawn_file_meta_reencrypt, and private helpers
  ReencryptOutcome/resolve_and_fetch_file_meta to metadata.rs
- Move CipherBoxFS struct, inherent impl, uuid_from_ino, and mount_point to fs.rs
- Move T-45-08 merge tests into metadata.rs mod tests
- Add cfg-gated pub mod metadata/fs declarations and re-exports to lib.rs
- No visibility changes; no public API changes
